all: Use Guava Joiner and Range rather than custom classes
There is no reason to reimplement something we can get from Guava.

Requires synchronous upgrade of non-pool components.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>

+++ modules/dcache/src/main/java/org/dcache/chimera/nfsv41/mover/NFSv4MoverHandler.java
@@ -53,7 +53,7 @@ public class NFSv4MoverHandler {
         ServerIdProvider idProvider =  HimeraNFS4Utils.cellNameToServerIdProvider(serverId);
         _embededDS = new NFSServerV41(_operationFactory, null, null, _fs, new SimpleIdMap(), null, idProvider);
         _rpcService = new OncRpcSvc(
-                new org.glassfish.grizzly.PortRange((int)portRange.getLower(), (int)portRange.getUpper()),
+                new org.glassfish.grizzly.PortRange(portRange.getLower(), portRange.getUpper()),
                 IpProtocolType.TCP, false);
 
         final Map<OncRpcProgram, RpcDispatchable> programs = new HashMap<OncRpcProgram, RpcDispatchable>();
